requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
pkg_resources==0.0.0
prompt-toolkit==3.0.20
python-magic==0.4.24
pytz==2021.3
redis==3.5.3
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
